Downgrade to Java 8 as source baseline
This was a cheap change and might get some more users from JBang since
JBang probably has lots of people not being able to run Java 11 scripts.

Also, I removed the single occurrence of Java 8 Streams because a
for-loop fits better with the rest of the -- super simple -- code style.

Closes #3.