Improve number parsing
This aligns JSON Wheel's number parsing closer with the JSON spec in
that it supports e-Notation, _all_ allowed symbols ('E', 'e', '.', '-',
'+') and arbitrary precision.

Note: The parser still does not care whether the above symbols appear in
correct order. It tries to determine the end of the number literal and
then parse it into a `java.lang.Number` type.

Closes #2.